Update rayleigh_selection.R
When checking the number of points in the complex against the number of points over which the function is defined, there was a bug in the edge case where nerve_complexes would remove vertices containing exactly one point. The complex would then have fewer points than f, which is the intended behavior. However, rayleigh_selection would terminate. This has been updated so that the program stops iff there are points in the complex for which f is not defined (i.e., the largest point in the complex is larger than the number of points over which f is defined).